requirements.txt
ultralytics
torch
torchvision
opencv-python
numpy
Pillow
